Refactor: 기존에 index.tsx 에 모여있던, field 컴포넌트들을 파일로 분리
