+++ app/recruitment/new/_components/fields/index.tsx
@@ -1,0 +1,26 @@
+'use client';
+import { CalendarField } from '~/app/recruitment/new/_components/fields/calendar-field';
+import { ImageField } from '~/app/recruitment/new/_components/fields/image-field';
+import { InputField } from '~/app/recruitment/new/_components/fields/input-field';
+import {
+  RadioGroupField,
+  RadioGroupFieldItem,
+} from '~/app/recruitment/new/_components/fields/radio-field';
+import {
+  SelectField,
+  SelectFieldItem,
+} from '~/app/recruitment/new/_components/fields/select-field';
+import { SliderField } from '~/app/recruitment/new/_components/fields/slider-field';
+import { TextareaField } from '~/app/recruitment/new/_components/fields/textarea-field';
+
+export {
+  CalendarField,
+  InputField,
+  RadioGroupField,
+  RadioGroupFieldItem,
+  SelectField,
+  SelectFieldItem,
+  SliderField,
+  TextareaField,
+  ImageField,
+};
